+++ packages/localizations/src/cs-CZ.ts
@@ -71,6 +71,7 @@ export const csCZ: LocalizationResource = {
     manageSubscription: undefined,
     month: undefined,
     monthly: undefined,
+    pastDue: undefined,
     pay: undefined,
     paymentMethods: undefined,
     paymentSource: {
@@ -119,6 +120,7 @@ export const csCZ: LocalizationResource = {
     sameDay: "Dnes v  {{ date | timeString('cs-CZ') }}",
   },
   dividerText: 'nebo',
+  footerActionLink__alternativePhoneCodeProvider: undefined,
   footerActionLink__useAnotherMethod: 'Použít jinou metodu',
   footerPageLink__help: 'Nápověda',
   footerPageLink__privacy: 'Ochrana soukromí',