+++ packages/localizations/src/id-ID.ts
@@ -71,6 +71,7 @@ export const idID: LocalizationResource = {
     manageSubscription: undefined,
     month: undefined,
     monthly: undefined,
+    pastDue: undefined,
     pay: undefined,
     paymentMethods: undefined,
     paymentSource: {
@@ -119,6 +120,7 @@ export const idID: LocalizationResource = {
     sameDay: "Hari ini pada {{ date | timeString('id-ID') }}",
   },
   dividerText: 'atau',
+  footerActionLink__alternativePhoneCodeProvider: undefined,
   footerActionLink__useAnotherMethod: 'Gunakan metode lain',
   footerPageLink__help: 'Bantuan',
   footerPageLink__privacy: 'Privasi',