+++ packages/remix/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.7.7
+
+### Patch Changes
+
+- Updated dependencies [[`6ed3dfc`](https://github.com/clerk/javascript/commit/6ed3dfc1bc742ac9d9a2307fe8e4733411cbc0d7), [`22c3363`](https://github.com/clerk/javascript/commit/22c33631f7f54b4f2179bf16f548fee1a237976e), [`ac6b231`](https://github.com/clerk/javascript/commit/ac6b23147e5e0aa21690cc20a109ed9a8c8f6e5b)]:
+  - @clerk/types@4.59.2
+  - @clerk/backend@1.33.1
+  - @clerk/clerk-react@5.31.7
+  - @clerk/shared@3.9.4
+
 ## 4.7.6
 
 ### Patch Changes